Add util file + new constants + minor code style fixes
