+++ TheMovieDb/Constants.swift
@@ -7,9 +7,10 @@
 
 import Foundation
 
-struct Constant  { //constants
+struct Constants  { //constants
     static let cellIdentifier: String = "ReusableCell"
     static let cellXibName: String = "MovieCell"
+    static let segueIdentifier: String = "showDetail"
     
     struct MovieLaunch {
         static let nowPlaying: String = "now_playing"
@@ -21,5 +22,7 @@ struct Constant  { //constants
     
     struct URLS{
         static let imageURL: String = "https://image.tmdb.org/t/p/w500"
+        static let movieBaseURL: String = "https://api.themoviedb.org/3"
+        static let apiKey: String = "444cd656b00475d785aa41a9c43b2e44"
     }
 }
