+++ TheMovieDb/Extensions + Utils/Extensions.swift
@@ -32,25 +32,4 @@ public class MockURLSession: URLSession {
     }
 }
 
-extension UIImageView {
 
-func posterFetcher(url: URL, contentMode mode:  UIView.ContentMode = .scaleAspectFit) {
-
-    URLSession.shared.dataTask(with: url) { data, response, error in
-        guard
-            let httpURLResponse = response as? HTTPURLResponse,
-                httpURLResponse.statusCode == 200,
-            let mimeType = response?.mimeType, mimeType.hasPrefix("image"),
-            let data = data, error == nil,
-            let image = UIImage(data: data)
-            else {
-                return
-                }
-            DispatchQueue.main.async() {
-                self.image = image
-                
-            }
-        }.resume()
-    }
-
-}
