+++ TheMovieDb/Models/MovieManager.swift
@@ -8,18 +8,16 @@
 import Foundation
 
 class MovieManager{
-    let movieBaseURL = "https://api.themoviedb.org/3"
-    let apiKey = "?api_key=444cd656b00475d785aa41a9c43b2e44"
     var apiService = APIService()
     var movies = [Movie]()
     
     func fetchMoviesData(type: String, completion: @escaping () -> ()) {
         let urlString: String
         
-        if type == "trending" {
-            urlString = "\(movieBaseURL)/\(type)/movie/day\(apiKey)&language=en"
+        if type == Constants.MovieLaunch.trending {
+            urlString = "\(Constants.URLS.movieBaseURL)/\(type)/movie/day?api_key=\(Constants.URLS.apiKey)&language=en"
         } else {
-            urlString = "\(movieBaseURL)/movie/\(type)\(apiKey)&language=en"
+            urlString = "\(Constants.URLS.movieBaseURL)/movie/\(type)?api_key=\(Constants.URLS.apiKey)&language=en"
         }
         // weak self - prevent retain cycles
         apiService.getMoviesData(url: urlString) { [weak self] (result) in
