+++ TheMovieDb/Views/MovieCell.swift
@@ -18,30 +18,6 @@ class MovieCell: UITableViewCell {
     
     private var urlString: String = ""
     
-    func showStar(value: Int) -> String {
-        var star:String = ""
-        if value < 20 && value >= 0 {
-            star = "★☆☆☆☆"
-        }
-        else if value < 40 {
-            star = "★★☆☆☆"
-        }
-        else if value < 50 {
-            star = "★★★☆☆"
-        }
-        else if value < 70 {
-            star = "★★★★☆"
-        }
-        else if value < 90 && value <= 100 {
-            star = "★★★★★"
-        }
-        else {
-            star = "TBD"
-        }
-        return star;
-        
-    }
-    
     func setCellWithValuesOf(_ movie:Movie){
         updateUI(title: movie.originalTitle, releaseDate: movie.releaseDate, rating: movie.voteAverage, overview: movie.overview, poster: movie.posterPath)
     }
@@ -49,31 +25,20 @@ class MovieCell: UITableViewCell {
     
     private func updateUI(title: String?, releaseDate: String?, rating: Double?, overview: String?, poster: String?){
         self.movieTitleLabel.text = title
+        self.movieTitleLabel.adjustsFontSizeToFitWidth = true
+
         self.releaseDateLabel.text = fixedDateFormatter(releaseDate)
         guard let average = rating else { return }
         self.ratingLabel.text =  showStar(value: Int(average * 10))
         self.movieOverviewLabel.text = overview
         
         guard let posterPath = poster else { return }
-        urlString = "\(Constant.URLS.imageURL)\(posterPath)"
+        urlString = "\(Constants.URLS.imageURL)\(posterPath)"
         
         if let imageURL = URL(string: urlString){
-            movieImageView.posterFetcher(url: imageURL)
+            movieImageView.imageFetcher(url: imageURL)
         }
         
-    }
-    
-    func fixedDateFormatter(_ date: String?) -> String {
-        var fixDate: String = ""
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyy-MM-dd"
-        if let originalDate = date {
-            if let newDate = dateFormatter.date(from: originalDate) {
-                dateFormatter.dateFormat = "dd.MM.yyyy"
-                fixDate = dateFormatter.string(from: newDate)
-            }
-        }
-        return fixDate
     }
     
 }
